requirements.txt
dbt-postgres
dbt-airflow
dbt-core
click
snowflake-connector-python